+++ src/main/java/com/webapps2022/jsf/CurrencyTransactionBean.java
@@ -8,7 +8,7 @@ import com.webapps2022.ejb.CurrencyTransactionService;
 import com.webapps2022.entity.CurrencyEnum;
 import com.webapps2022.entity.CurrencyTransaction;
 import com.webapps2022.entity.SystemUser;
-import java.math.BigDecimal;
+
 import java.util.List;
 import javax.ejb.EJB;
 import javax.enterprise.context.RequestScoped;
@@ -28,7 +28,7 @@ public class CurrencyTransactionBean {
     CurrencyTransactionService transactionService;
 
     private String otherUsername;
-    private BigDecimal currencyCount;
+    private Float currencyCount;
 
     public CurrencyTransactionService getTransactionService() {
         return transactionService;
@@ -46,11 +46,11 @@ public class CurrencyTransactionBean {
         this.otherUsername = otherUsername;
     }
 
-    public BigDecimal getCurrencyCount() {
+    public Float getCurrencyCount() {
         return currencyCount;
     }
 
-    public void setCurrencyCount(BigDecimal currencyCount) {
+    public void setCurrencyCount(Float currencyCount) {
         this.currencyCount = currencyCount;
     }
